back/src/main/java/com/studypot/back/applications/UserService.java
package com.studypot.back.applications;

import com.studypot.back.domain.User;
import com.studypot.back.domain.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class UserService {

  private UserRepository userRepository;

  @Autowired
  public UserService(UserRepository userRepository) {
    this.userRepository = userRepository;
  }

  public User registerUser(String name, String email, String password) {

    User user = User.builder()
        .name(name)
        .email(email)
        .password(password)
        .build();

    return userRepository.save(user);
  }
}
